+++ src/django_auth_lti/const.py
@@ -1,0 +1,10 @@
+# LTI role name constants
+
+ADMINISTRATOR = 'urn:lti:instrole:ims/lis/Administrator'
+LEARNER = 'Learner'
+INSTRUCTOR = 'Instructor'
+CONTENT_DEVELOPER = 'ContentDeveloper'
+OBSERVER = 'urn:lti:instrole:ims/lis/Observer'
+TEACHING_ASSISTANT = 'urn:lti:role:ims/lis/TeachingAssistant'
+
+TEACHING_STAFF_ROLES = [ADMINISTRATOR, INSTRUCTOR, TEACHING_ASSISTANT, CONTENT_DEVELOPER]